Implement a Species model
This commit implements a Species model that has the following fields [datatypes]:
  common_name [string]
  scientific_name [string]

The Species model validates for the presence of each field before creating a new record.

This commit also removes a sort of flimsy spec whose intention was to check the type of the values for the Species fields. It did this by trying to input an integer instead of a string, and expecting that the record would not be created. This didn't work because it turns out that ActiveRecord is saving the integer as a string anyway (for example, "3" instead of 3). It doesn't look like ActiveRecord has a way to validate that a field value should be a string. In the interest of time, I'm going to move on without validating the types of the field values.