Specify responding to an invalid POST request
When an HTTP client sends a POST request with query parameters that cannot be used to create a valid Observation record, the server should not create a new Observation record in the database and should send a 400 Bad Request response.

This commit also changes the RSpec matchers in the rest of the request specs to explicitly expect a 200 or 400 response code. This allows more distinct differentiation between different HTTP responses versus the previous usage of the RSpec be_success matcher.