Improve readability of a spec's request parameters
This commit changes the POST request in the Observation request specification for better readability.

Previously, the query parameters of the POST requests were hard-coded into a complete query string, which was long and difficult to read. The query parameters are now set using the params: {} option of the ActionDispatch::Integration::Session object as described here: http://api.rubyonrails.org/classes/ActionDispatch/Integration/Session.html#method-i-process. This makes the structure of the query parameters more visually clear.