Implement authorization in HTTP requests
An HTTP request responds to any unauthorized request with a 400 and a blank response body.

In order for a request to be considered authorized, the request must have an Authorization header with the value 'Bearer <AUTH_TOKEN>', where <AUTH_TOKEN> matches an authorization token that is saved as an environment variable.