Add more test cases
This commit specifies that a request with an authorization header containing the correct authorization token but otherwise improperly structured should result in a 400 response. In this case, the improper structure involves a typo in the 'Bearer' component of the authorization header's value.

This commit also adds all of the test cases to the POST request section that previously were only tested for GET requests. Specifically, POST and GET requests are now tested for:
  - an authorization header that is properly structured but has the wrong authorization token
  - an authorization header that has the correct authorization token but is improperly structured
  - no authorization header

All of these cases are already green under the current implementation.