Eliminate some repetition in the request specs
This commit makes the specs slightly less verbose by moving a repeated line of setup code into a single line of code in a setup block.

More specifically, the authorization token was previously set as an environment variable in each test case individually, and is now set in a block that runs before each test case.