Pivot to make the data public
This commit represents a change where the server will respond to all GET requests with the requested data, in the aim of intentionally making the data publicly available. With these changes, only POST requests require authentication.

This represents a tradeoff: the open data may provide some benefit to society by providing an open knowledge resource for understanding the locations of biological species in the natural world, but it opens up a security vulnerability by enabling an attacker to send lots of GET requests that each result in a response containing data. This could make the production application more costly to operate and possibly disable the application if an attacker drives up the operating bills enough. Is it worth it? Time will tell.