Specify an observation's species in the response
This commit specifies that the server should respond to a request for observations with not only the observations themselves, but also the species associated with those observations.

Each observation belongs to a species. When an observation is requested, the common name and scientific name of the observation's species should be sent in the response, in addition to the id, latitude, longitude, and date of the observation itself.

Also, in order to keep a consistent data exchange format, every data tuple in the response should be an object in a collection (even if only one tuple is requested and the collection has a length of 1). Specifically, if a single observation is requested, the response should be a collection containing a single entry for the requested observation and its species. If all observations are requested, the response should be a collection containing entries that each contain an observation and its species.